crispy forms added

+++ src/blog/views.py
@@ -1,4 +1,5 @@
 from multiprocessing import context
+from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
 from blog.forms import CommentForm, PostForm
 from .models import Like, Post
@@ -48,6 +49,9 @@ def post_detail(request, slug):
 
 def post_update(request, slug):
     obj = get_object_or_404(Post, slug=slug)
+
+    if request.user.id != obj.author.id:
+        return HttpResponse("You are not authorized")
     form = PostForm(request.POST or None, request.FILES or None, instance=obj)
     if form.is_valid():
         form.save()
@@ -62,6 +66,9 @@ def post_update(request, slug):
 
 def post_delete(request, slug):
     obj = get_object_or_404(Post, slug=slug)
+
+    if request.user.id != obj.author.id:
+        return HttpResponse("You are not authorized")
     if request.method == 'POST':
         obj.delete()
         return redirect("blog:list")
